requirements.txt
Cython
#insightface==0.6.2
#mmcv-full==1.3.3
numpy
onnxruntime
opencv-python-headless
openmim
torch
torchvision
gradio
mmdet
mmcv-full